fix repo,model,controller,service

+++ src/main/java/su/oilmap/oilmap/service/StationService.java
@@ -5,5 +5,5 @@ import su.oilmap.oilmap.domain.Station;
 import java.util.Optional;
 
 public interface StationService {
-    Optional<Station> getStationById(int id);
+    Optional<Station> getStationById(long id);
 }
